moved VectorTileClipper to new module
